everything makes on my upgraded VM; had to do a few casts; still working on desktop..

+++ src/edb/Connect.cc
@@ -144,7 +144,7 @@ DBResult::unpack()
         return new ResType();
     }
     cerr << "b\n";
-    size_t rows = mysql_num_rows(n);
+    size_t rows = (size_t)mysql_num_rows(n);
     int cols  = -1;
     if (rows > 0) {
         cols = mysql_num_fields(n);
